+++ lib/onlyoffice_webdriver_wrapper/webdriver/webdriver_screenshot_helper.rb
@@ -4,6 +4,12 @@ require 'onlyoffice_s3_wrapper'
 module OnlyofficeWebdriverWrapper
   # Working with screenshots
   module WebdriverScreenshotHelper
+    # @return [OnlyofficeS3Wrapper::AmazonS3Wrapper] s3 wrapper
+    def amazon_s3_wrapper
+      @amazon_s3_wrapper ||= OnlyofficeS3Wrapper::AmazonS3Wrapper.new(bucket_name: 'webdriver-wrapper-screenshots',
+                                                                      region: 'us-east-1')
+    end
+
     # @return [String] location of screenshot to take
     def screenshot_folder
       '/tmp/screenshot/WebdriverError'
@@ -12,7 +18,7 @@ module OnlyofficeWebdriverWrapper
     def get_screenshot_and_upload(path_to_screenshot = "#{screenshot_folder}/#{StringHelper.generate_random_string}.png")
       begin
         get_screenshot(path_to_screenshot)
-        cloud_screenshot = OnlyofficeS3Wrapper::AmazonS3Wrapper.new.upload_file_and_make_public(path_to_screenshot, 'screenshots')
+        cloud_screenshot = amazon_s3_wrapper.upload_file_and_make_public(path_to_screenshot)
         File.delete(path_to_screenshot) if File.exist?(path_to_screenshot)
         OnlyofficeLoggerHelper.log("upload screenshot: #{cloud_screenshot}")
         return cloud_screenshot
@@ -43,14 +49,14 @@ module OnlyofficeWebdriverWrapper
         if @headless.headless_instance.nil?
           system_screenshot("/tmp/#{screenshot_name}.png")
           begin
-            link = OnlyofficeS3Wrapper::AmazonS3Wrapper.new.upload_file_and_make_public("/tmp/#{screenshot_name}.png", 'screenshots')
+            link = amazon_s3_wrapper.upload_file_and_make_public("/tmp/#{screenshot_name}.png")
           rescue Exception => e
             OnlyofficeLoggerHelper.log("Error in get screenshot: #{e}. System screenshot #{link}")
           end
         else
           @headless.take_screenshot("/tmp/#{screenshot_name}.png")
           begin
-            link = OnlyofficeS3Wrapper::AmazonS3Wrapper.new.upload_file_and_make_public("/tmp/#{screenshot_name}.png", 'screenshots')
+            link = amazon_s3_wrapper.upload_file_and_make_public("/tmp/#{screenshot_name}.png")
           rescue Exception => e
             OnlyofficeLoggerHelper.log("Error in get screenshot: #{e}. Headless screenshot #{link}")
           end